Move eere calculations into seperate file, and move calculating top percent generation out of calculateEere() function
