+++ client/src/app/calculations/transportation.ts
@@ -1,7 +1,4 @@
-import {
-  RegionalLoadData,
-  EEREDefaultData,
-} from 'app/redux/reducers/geography';
+import { RegionalLoadData } from 'app/redux/reducers/geography';
 import type { GeographicFocus } from 'app/redux/reducers/geography';
 import type {
   RegionalScalingFactors,
@@ -204,10 +201,6 @@ export type TotalYearlyEmissionChanges = ReturnType<
 >;
 export type TotalYearlyEVEnergyUsage = ReturnType<
   typeof calculateTotalYearlyEVEnergyUsage
->;
-export type HourlyEVLoad = ReturnType<typeof calculateHourlyEVLoad>;
-export type HourlyRenewableEnergyProfile = ReturnType<
-  typeof calculateHourlyRenewableEnergyProfile
 >;
 export type VehicleSalesAndStock = ReturnType<
   typeof calculateVehicleSalesAndStock
@@ -1575,106 +1568,6 @@ export function calculateTotalYearlyEmissionChanges(
     },
     { CO2: 0, NOX: 0, SO2: 0, PM25: 0, VOCs: 0, NH3: 0 },
   );
-
-  return result;
-}
-
-/**
- * Hourly EV load.
- *
- * Excel: Data in column Y of the "CalculateEERE" sheet (Y5:Y8788).
- */
-export function calculateHourlyEVLoad(options: {
-  regionalLoad: RegionalLoadData[];
-  dailyStats: DailyStats;
-  hourlyEVChargingPercentages: HourlyEVChargingPercentages;
-  monthlyDailyEVEnergyUsage: MonthlyDailyEVEnergyUsage;
-}) {
-  const {
-    regionalLoad,
-    dailyStats,
-    hourlyEVChargingPercentages,
-    monthlyDailyEVEnergyUsage,
-  } = options;
-
-  if (
-    regionalLoad.length === 0 ||
-    Object.keys(dailyStats).length === 0 ||
-    Object.keys(hourlyEVChargingPercentages).length === 0 ||
-    Object.keys(monthlyDailyEVEnergyUsage).length === 0
-  ) {
-    return [];
-  }
-
-  const result = regionalLoad.map((data) => {
-    if (
-      !data.hasOwnProperty('hour') &&
-      !data.hasOwnProperty('day') &&
-      !data.hasOwnProperty('month')
-    ) {
-      return 0;
-    }
-
-    // NOTE: `regionalLoad` data's hour value is zero indexed, so to match it
-    // with the hours stored as keys in `hourlyEVChargingPercentages`, we need
-    // to add 1 to `regionalLoad` data's hour value
-    const hour = data.hour + 1;
-    const day = data.day;
-    const month = data.month;
-
-    const evChargingPercentage = hourlyEVChargingPercentages[hour];
-    const dayTypeField = dailyStats[month][day].isWeekend
-      ? 'weekend'
-      : 'weekday';
-
-    const evLoad =
-      evChargingPercentage.batteryEVs[dayTypeField] *
-        monthlyDailyEVEnergyUsage[month].batteryEVs[dayTypeField] +
-      evChargingPercentage.hybridEVs[dayTypeField] *
-        monthlyDailyEVEnergyUsage[month].hybridEVs[dayTypeField] +
-      evChargingPercentage.transitBuses[dayTypeField] *
-        monthlyDailyEVEnergyUsage[month].transitBuses[dayTypeField] +
-      evChargingPercentage.schoolBuses[dayTypeField] *
-        monthlyDailyEVEnergyUsage[month].schoolBuses[dayTypeField];
-
-    return evLoad;
-  });
-
-  return result;
-}
-
-/**
- * Hourly Renewable Energy Profile.
- *
- * Excel: Data in column H of the "CalculateEERE" sheet (H5:H8788).
- */
-export function calculateHourlyRenewableEnergyProfile(options: {
-  eereDefaults: EEREDefaultData[];
-  lineLoss: number;
-  onshoreWind: number;
-  offshoreWind: number;
-  utilitySolar: number;
-  rooftopSolar: number;
-}) {
-  const {
-    eereDefaults,
-    lineLoss,
-    onshoreWind: onshoreWindInput,
-    offshoreWind: offshoreWindInput,
-    utilitySolar: utilitySolarInput,
-    rooftopSolar: rooftopSolarInput,
-  } = options;
-
-  if (eereDefaults.length === 0) return [];
-
-  const result = eereDefaults.map((data) => {
-    const onshoreWind = onshoreWindInput * data.onshore_wind;
-    const offshoreWind = offshoreWindInput * (data.offshore_wind || 0);
-    const utilitySolar = utilitySolarInput * data.utility_pv;
-    const rooftopSolar = (rooftopSolarInput * data.rooftop_pv) / (1 - lineLoss);
-
-    return -1 * (onshoreWind + offshoreWind + utilitySolar + rooftopSolar);
-  });
 
   return result;
 }
